47/51 Links da Pagina inicial
Links da Pagina inicial

+++ meu-anuncio/app/Http/Controllers/Admin/AnuncioController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 use App\Models\Anuncio;
@@ -12,14 +13,30 @@ use App\Models\Municipio;
 
 class AnuncioController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
+        if (!Auth::user()->can('listar-anuncios')) {
+            $request->session()->flash(
+                'mensagem',
+                ['msg' => 'Erro: Sem acesso à funcionalidade!', 'class' => 'red white-text']
+            );
+            return redirect()->route('admin.home');
+        }
+
         $anuncios = Anuncio::all();
         return view('admin.anuncios.index', compact('anuncios'));
     }
 
-    public function cadastrar()
+    public function cadastrar(Request $request)
     {
+        if (!Auth::user()->can('cadastrar-anuncios')) {
+            $request->session()->flash(
+                'mensagem',
+                ['msg' => 'Erro: Sem acesso à funcionalidade!', 'class' => 'red white-text']
+            );
+            return redirect()->route('admin.home');
+        }
+
         $categorias = Categoria::all()->sortBy('titulo');
         $municipios = Municipio::all()->sortBy('nome');
         return view('admin.anuncios.cadastrar', compact(['categorias', 'municipios']));
@@ -57,6 +74,14 @@ class AnuncioController extends Controller
 
     public function salvar(Request $request)
     {
+        if (!Auth::user()->can('cadastrar-anuncios')) {
+            $request->session()->flash(
+                'mensagem',
+                ['msg' => 'Erro: Sem acesso à funcionalidade!', 'class' => 'red white-text']
+            );
+            return redirect()->route('admin.home');
+        }
+
         $dados = $request->all();
 
         $anuncio = new Anuncio();
@@ -70,8 +95,16 @@ class AnuncioController extends Controller
         return redirect()->route('admin.anuncios');
     }
 
-    public function alterar($id)
+    public function alterar(Request $request, $id)
     {
+        if (!Auth::user()->can('atualizar-anuncios')) {
+            $request->session()->flash(
+                'mensagem',
+                ['msg' => 'Erro: Sem acesso à funcionalidade!', 'class' => 'red white-text']
+            );
+            return redirect()->route('admin.home');
+        }
+
         $anuncio = Anuncio::find($id);
         $categorias = Categoria::all()->sortBy('titulo');
         $municipios = Municipio::all()->sortBy('nome');
@@ -80,6 +113,14 @@ class AnuncioController extends Controller
 
     public function atualizar(Request $request, $id)
     {
+        if (!Auth::user()->can('atualizar-anuncios')) {
+            $request->session()->flash(
+                'mensagem',
+                ['msg' => 'Erro: Sem acesso à funcionalidade!', 'class' => 'red white-text']
+            );
+            return redirect()->route('admin.home');
+        }
+
         $dados = $request->all();
 
         $anuncio = Anuncio::find($id);
@@ -95,12 +136,22 @@ class AnuncioController extends Controller
 
     public function remover(Request $request, $id)
     {
+        if (!Auth::user()->can('remover-anuncios')) {
+            $request->session()->flash(
+                'mensagem',
+                ['msg' => 'Erro: Sem acesso à funcionalidade!', 'class' => 'red white-text']
+            );
+            return redirect()->route('admin.home');
+        }
+
         $anuncio = Anuncio::find($id);
         $anuncio->imagens()->delete();
         $anuncio->delete();
 
-        $request->session()->flash('mensagem',
-            ['msg' => 'Anúncio removido com sucesso!', 'class' => 'green white-text']);
+        $request->session()->flash(
+            'mensagem',
+            ['msg' => 'Anúncio removido com sucesso!', 'class' => 'green white-text']
+        );
         return redirect()->route('admin.anuncios');
     }
 }
